Further splitting
o FAT16 related NAND functions are now in nandfat.c / .h
o This makes sense because more file dumping / injection options are
coming
o Also renamed GetNandPartition() to GetPartitionInfo()
