+++ source/decryptor/nand.c
@@ -90,7 +90,7 @@ static inline int WriteNandSectors(u32 sector_no, u32 numsectors, u8 *in)
     } else return sdmmc_nand_writesectors(sector_no, numsectors, in);
 }
 
-PartitionInfo* GetNandPartition(u32 partition_id)
+PartitionInfo* GetPartitionInfo(u32 partition_id)
 {
     if (partition_id == P_CTRNAND)
         return &(partitions[(GetUnitPlatform() == PLATFORM_3DS) ? 5 : 6]);
@@ -192,100 +192,6 @@ u32 GetNandCtr(u8* ctr, u32 offset)
 
     return 0;
 }
-
-u32 SeekFileInNand(u32* offset, u32* size, const char* path, PartitionInfo* partition)
-{
-    // poor mans NAND FAT file seeker:
-    // - path must be in FAT 8+3 format, without dots or slashes
-    //   example: DIR1_______DIR2_______FILENAMEEXT
-    // - can't handle long filenames
-    // - dirs must not exceed 1024 entries
-    // - fragmentation not supported
-    
-    const static char zeroes[8+3] = { 0x00 };
-    u8* buffer = BUFFER_ADDRESS;
-    u32 p_size = partition->size;
-    u32 p_offset = partition->offset;
-    
-    u32 cluster_size;
-    u32 cluster_start;
-    bool found = false;
-    
-    if (strnlen(path, 256) % (8+3) != 0)
-        return 1;
-    
-    DecryptNandToMem(buffer, p_offset, NAND_SECTOR_SIZE, partition);
-    
-    // good FAT header description found here: http://www.compuphase.com/mbr_fat.htm
-    u32 fat_start = NAND_SECTOR_SIZE * (*((u16*) (buffer + 0x0E)));
-    u32 fat_size = NAND_SECTOR_SIZE * (*((u16*) (buffer + 0x16)) * buffer[0x10]);
-    u32 root_size = *((u16*) (buffer + 0x11)) * 0x20;
-    cluster_start = fat_start + fat_size + root_size;
-    cluster_size = buffer[0x0D] * NAND_SECTOR_SIZE;
-    
-    for (*offset = p_offset + fat_start + fat_size; strnlen(path, 256) >= 8+3; path += 8+3) {
-        if (*offset - p_offset > p_size)
-            return 1;
-        found = false;
-        DecryptNandToMem(buffer, *offset, cluster_size, partition);
-        for (u32 i = 0x00; i < cluster_size; i += 0x20) {
-            // skip invisible, deleted and lfn entries
-            if ((buffer[i] == '.') || (buffer[i] == 0xE5) || (buffer[i+0x0B] == 0x0F))
-                continue;
-            else if (memcmp(buffer + i, zeroes, 8+3) == 0)
-                return 1;
-            u32 p; // search for path in fat folder structure, accept '?' wildcards
-            for (p = 0; (p < 8+3) && (path[p] == '?' || buffer[i+p] == path[p]); p++);
-            if (p != 8+3) continue;
-            // entry found, store offset and move on
-            *offset = p_offset + cluster_start + (*((u16*) (buffer + i + 0x1A)) - 2) * cluster_size;
-            *size = *((u32*) (buffer + i + 0x1C));
-            found = true;
-            break;
-        }
-        if (!found) break;
-    }
-    
-    return (found) ? 0 : 1;
-}
-
-u32 DumpTicket() {
-    PartitionInfo* ctrnand_info = GetNandPartition(P_CTRNAND);
-    u32 offset;
-    u32 size;
-    
-    Debug("Searching for ticket.db...");
-    if (SeekFileInNand(&offset, &size, "DBS        TICKET  DB ", ctrnand_info) != 0) {
-        Debug("Failed!");
-        return 1;
-    }
-    Debug("Found at %08X, size %uMB", offset, size / (1024 * 1024));
-    
-    if (DecryptNandToFile((IsEmuNand()) ? "/ticket_emu.db" : "/ticket.db", offset, size, ctrnand_info) != 0)
-        return 1;
-    
-    return 0;
-}
-
-u32 DumpSeedsave()
-{
-    PartitionInfo* ctrnand_info = GetNandPartition(P_CTRNAND);
-    u32 offset;
-    u32 size;
-    
-    Debug("Searching for seedsave...");
-    if (SeekFileInNand(&offset, &size, "DATA       ???????????SYSDATA    0001000F   00000000   ", ctrnand_info) != 0) {
-        Debug("Failed!");
-        return 1;
-    }
-    Debug("Found at %08X, size %ukB", offset, size / 1024);
-    
-    if (DecryptNandToFile("/seedsave.bin", offset, size, ctrnand_info) != 0)
-        return 1;
-    
-    return 0;
-}
-
 u32 DecryptNandToMem(u8* buffer, u32 offset, u32 size, PartitionInfo* partition)
 {
     CryptBufferInfo info = {.keyslot = partition->keyslot, .setKeyY = 0, .size = size, .buffer = buffer, .mode = partition->mode};
@@ -373,17 +279,17 @@ u32 DecryptAllNandPartitions() {
     u32 result = 0;
     
     for (u32 partition_id = 0; partition_id < 6; partition_id++)
-        result |= DecryptNandPartition(GetNandPartition(partition_id));
+        result |= DecryptNandPartition(GetPartitionInfo(partition_id));
     
     return result;
 }
 
 u32 DecryptTwlNandPartition() {
-    return DecryptNandPartition(GetNandPartition(P_TWLN)); // TWLN
+    return DecryptNandPartition(GetPartitionInfo(P_TWLN)); // TWLN
 }
     
 u32 DecryptCtrNandPartition() {
-    return DecryptNandPartition(GetNandPartition(P_CTRNAND)); // CTRNAND O3DS / N3DS
+    return DecryptNandPartition(GetPartitionInfo(P_CTRNAND)); // CTRNAND O3DS / N3DS
 }
 
 u32 EncryptMemToNand(u8* buffer, u32 offset, u32 size, PartitionInfo* partition)
@@ -505,15 +411,15 @@ u32 InjectAllNandPartitions() {
     u32 result = 1;
     
     for (u32 partition_id = 0; partition_id < 6; partition_id++)
-        result &= InjectNandPartition(GetNandPartition(partition_id));
+        result &= InjectNandPartition(GetPartitionInfo(partition_id));
     
     return result;
 }
 
 u32 InjectTwlNandPartition() {
-    return InjectNandPartition(GetNandPartition(P_TWLN)); // TWLN
+    return InjectNandPartition(GetPartitionInfo(P_TWLN)); // TWLN
 }
 
 u32 InjectCtrNandPartition() {
-    return InjectNandPartition(GetNandPartition(P_CTRNAND)); // CTRNAND O3DS / N3DS
+    return InjectNandPartition(GetPartitionInfo(P_CTRNAND)); // CTRNAND O3DS / N3DS
 }
